backend/app/identification/model.py
import numpy as np
from PIL import Image
import tensorflow as tf
import io

# Path to your trained CNN model
#MODEL_PATH = "path/to/your/model"

# In app/identification/model.py
class PlantIdentifier:
    def __init__(self):
        # Comment out model loading for now
        # self.model = tf.keras.models.load_model(MODEL_PATH)
        self.model = None
        self.class_names = [
            "plant_type_1",
            "plant_type_2",
            # Add all plant types your model can identify
        ]
        
    def identify(self, image_bytes):
        # Return mock data for now
        return {
            "plant_name": "Sample Plant",  # Make sure this key matches what your frontend expects
            "probability": 95.0,           # Use probability instead of confidence for consistency
            "other_candidates": [           # Make sure this matches what the frontend expects
                {"plant_name": "Sample Plant", "probability": 95.0},
                {"plant_name": "Another Plant", "probability": 5.0}
            ]
        }

# Create a singleton instance
plant_identifier = PlantIdentifier()