backend/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0
python-jose==3.3.0
passlib==1.7.4
pymongo==4.5.0
python-multipart==0.0.6
python-dotenv==1.0.0
bcrypt==4.0.1
Pillow==10.0.1
tensorflow==2.19.0  # Adjust based on your CNN model requirements
numpy==1.24.3